requirements.txt
Flask
Flask_SQLAlchemy
python-dotenv
google-generativeai
SpeechRecognition
Pillow
